refactor: move collection defs out of core plugin
