--- src/core-services/core/util/layouts.js
@@ -1,3 +1,0 @@
-const layouts = [];
-
-export default layouts;
