+++ src/core-services/shop/index.js
@@ -14,6 +14,27 @@ export default async function register(app) {
   await app.registerPlugin({
     label: "Shop",
     name: "reaction-shop",
+    collections: {
+      Packages: {
+        name: "Packages",
+        indexes: [
+          // Create indexes. We set specific names for backwards compatibility
+          // with indexes created by the aldeed:schema-index Meteor package.
+          [{ name: 1, shopId: 1 }],
+          [{ "registry.provides": 1 }, { name: "c2_registry.$.provides" }]
+        ]
+      },
+      Shops: {
+        name: "Shops",
+        indexes: [
+          // Create indexes. We set specific names for backwards compatibility
+          // with indexes created by the aldeed:schema-index Meteor package.
+          [{ domains: 1 }, { name: "c2_domains" }],
+          [{ name: 1 }, { name: "c2_name" }],
+          [{ slug: 1 }, { name: "c2_slug", sparse: true, unique: true }]
+        ]
+      }
+    },
     graphQL: {
       resolvers,
       schemas
